refactor: move broker config to prover package
