+++ yarn-project/aztec/src/cli/aztec_start_options.ts
@@ -2,12 +2,7 @@ import { type ArchiverConfig, archiverConfigMappings } from '@aztec/archiver/con
 import { faucetConfigMapping } from '@aztec/aztec-faucet/config';
 import { sequencerClientConfigMappings } from '@aztec/aztec-node/config';
 import { botConfigMappings } from '@aztec/bot/config';
-import {
-  type ProverAgentConfig,
-  type ProverBrokerConfig,
-  proverAgentConfigMappings,
-  proverBrokerConfigMappings,
-} from '@aztec/circuit-types/config';
+import { type ProverAgentConfig, proverAgentConfigMappings } from '@aztec/circuit-types/config';
 import {
   type ConfigMapping,
   type EnvVar,
@@ -17,6 +12,7 @@ import {
 } from '@aztec/foundation/config';
 import { bootnodeConfigMappings, p2pConfigMappings } from '@aztec/p2p/config';
 import { proofVerifierConfigMappings } from '@aztec/proof-verifier/config';
+import { type ProverBrokerConfig, proverBrokerConfigMappings } from '@aztec/prover-client/broker';
 import { proverNodeConfigMappings } from '@aztec/prover-node/config';
 import { allPxeConfigMappings } from '@aztec/pxe/config';
 import { telemetryClientConfigMappings } from '@aztec/telemetry-client/start';
