+++ yarn-project/circuit-types/src/config.ts
@@ -1,3 +1,2 @@
 export { ProverAgentConfig, proverAgentConfigMappings } from './interfaces/prover-agent.js';
-export { ProverBrokerConfig, proverBrokerConfigMappings } from './interfaces/prover-broker.js';
 export { SequencerConfig, AllowedElement, SequencerConfigSchema } from './interfaces/configs.js';
