+++ yarn-project/prover-client/src/proving_broker/factory.ts
@@ -1,6 +1,6 @@
-import { type ProverBrokerConfig } from '@aztec/circuit-types';
 import { type TelemetryClient } from '@aztec/telemetry-client';
 
+import { type ProverBrokerConfig } from './config.js';
 import { ProvingBroker } from './proving_broker.js';
 import { InMemoryBrokerDatabase } from './proving_broker_database/memory.js';
 import { KVBrokerDatabase } from './proving_broker_database/persisted.js';
@@ -9,9 +9,7 @@ export async function createAndStartProvingBroker(
   config: ProverBrokerConfig,
   client: TelemetryClient,
 ): Promise<ProvingBroker> {
-  const database = config.proverBrokerDataDirectory
-    ? await KVBrokerDatabase.new(config, client)
-    : new InMemoryBrokerDatabase();
+  const database = config.dataDirectory ? await KVBrokerDatabase.new(config, client) : new InMemoryBrokerDatabase();
 
   const broker = new ProvingBroker(database, client, {
     jobTimeoutMs: config.proverBrokerJobTimeoutMs,
