+++ yarn-project/prover-client/src/proving_broker/index.ts
@@ -6,3 +6,4 @@ export * from './proving_broker_database/memory.js';
 export * from './proving_broker_database/persisted.js';
 export * from './proof_store.js';
 export * from './factory.js';
+export * from './config.js';
