+++ yarn-project/prover-client/src/proving_broker/proving_broker.test.ts
@@ -1,6 +1,5 @@
 import {
   type ProofUri,
-  type ProverBrokerConfig,
   type ProvingJob,
   type ProvingJobId,
   type ProvingJobStatus,
@@ -14,6 +13,7 @@ import { mkdtemp } from 'fs/promises';
 import { tmpdir } from 'os';
 import { join } from 'path';
 
+import { type ProverBrokerConfig } from './config.js';
 import { makeInputsUri, makeOutputsUri, makeRandomProvingJobId } from './fixtures.js';
 import { ProvingBroker } from './proving_broker.js';
 import { type ProvingBrokerDatabase } from './proving_broker_database.js';
@@ -25,8 +25,8 @@ describe.each([
   async () => {
     const directory = await mkdtemp(join(tmpdir(), 'proving-broker-test'));
     const config: ProverBrokerConfig = {
-      proverBrokerDataMapSizeKB: 1024 * 1024 * 1024, // 1GB
-      proverBrokerDataDirectory: directory,
+      dataStoreMapSizeKB: 1024 * 1024 * 1024, // 1GB
+      dataDirectory: directory,
       proverBrokerJobMaxRetries: 1,
       proverBrokerJobTimeoutMs: 1000,
       proverBrokerPollIntervalMs: 1000,
