+++ yarn-project/prover-node/src/config.ts
@@ -1,11 +1,6 @@
 import { type ArchiverConfig, archiverConfigMappings, getArchiverConfigFromEnv } from '@aztec/archiver/config';
 import { type ACVMConfig, type BBConfig } from '@aztec/bb-prover/config';
-import {
-  type ProverAgentConfig,
-  type ProverBrokerConfig,
-  proverAgentConfigMappings,
-  proverBrokerConfigMappings,
-} from '@aztec/circuit-types/config';
+import { type ProverAgentConfig, proverAgentConfigMappings } from '@aztec/circuit-types/config';
 import {
   type ConfigMappingsType,
   bigintConfigHelper,
@@ -14,6 +9,7 @@ import {
 } from '@aztec/foundation/config';
 import { type DataStoreConfig, dataConfigMappings, getDataConfigFromEnv } from '@aztec/kv-store/config';
 import { type P2PConfig, getP2PConfigFromEnv, p2pConfigMappings } from '@aztec/p2p/config';
+import { type ProverBrokerConfig, proverBrokerConfigMappings } from '@aztec/prover-client/broker';
 import {
   type ProverClientConfig,
   bbConfigMappings,
